refactor: Renamed getTimeRepresentationType and moved into its own util
