+++ src/screen-components/travel-details-screens/DepartureDetailsScreenComponent.tsx
@@ -594,7 +594,7 @@ function EstimatedCallRow({
             timeValues={{
               aimedTime: call.aimedDepartureTime,
               expectedTime: call.expectedDepartureTime,
-              missingRealTime: !call.realtime && isStartOfServiceJourney,
+              isRealtime: call.realtime || !isStartOfServiceJourney,
             }}
             roundingMethod="floor"
           />
