+++ src/screen-components/travel-details-screens/utils.ts
@@ -27,35 +27,6 @@ import {Statuses} from '@atb/theme';
 import {isDefined} from '@atb/utils/presence';
 import {EstimatedCallWithMetadata} from './use-departure-data';
 
-const DEFAULT_THRESHOLD_AIMED_EXPECTED_IN_MINUTES = 1;
-
-export type TimeValues = {
-  aimedTime: string;
-  expectedTime?: string;
-  missingRealTime?: boolean;
-};
-type TimeRepresentationType =
-  | 'no-realtime'
-  | 'no-significant-difference'
-  | 'significant-difference';
-
-export function getTimeRepresentationType({
-  missingRealTime,
-  aimedTime,
-  expectedTime,
-}: TimeValues): TimeRepresentationType {
-  if (missingRealTime) {
-    return 'no-realtime';
-  }
-  if (!expectedTime) {
-    return 'no-significant-difference';
-  }
-  const secondsDifference = Math.abs(secondsBetween(aimedTime, expectedTime));
-  return secondsDifference <= DEFAULT_THRESHOLD_AIMED_EXPECTED_IN_MINUTES * 60
-    ? 'no-significant-difference'
-    : 'significant-difference';
-}
-
 export const getNoticesForLeg = (leg: Leg) =>
   filterNotices([
     ...(leg.line?.notices || []),
@@ -511,6 +482,7 @@ export function getLineAndTimeA11yLabel(
 export function getNonFreeLegs(legs: Leg[]) {
   return legs.filter((leg) => !isFreeLeg(leg));
 }
+
 export function isFreeLeg(leg: Leg) {
   return (
     leg.mode === Mode.Foot ||
